fix(bundle): harden untrusted PE bounds checks against uint overflow
TryFindTrailingCertTableOffset parses the PE optional header's Security
data directory to locate a trailing Authenticode certificate table.
peOffset is read straight from the untrusted DOS header e_lfanew field
(fully attacker-controlled). The bounds check `peOffset + PeSignatureSize
+ CoffHeaderSize > fileLength` bound to the (uint, uint) operator overload
(the two operands are compile-time int constants, which C# implicitly
converts to uint), so a peOffset near uint.MaxValue could wrap the sum to
a small value and defeat the check.

Not exploitable today only because the resulting out-of-range Seek+Read
throws EndOfStreamException, which every caller (HasBundleFooter, Extract)
already catches and turns into a graceful failure -- but a security bounds
check on fully-untrusted bytes should be authoritative on its own, not
rely on an exception net downstream. Casting peOffset to long immediately
after the read makes every arithmetic expression derived from it long
for the rest of the method, so the check now rejects an out-of-range
offset directly instead of wrapping.

Added a reflection-based test (the method is intentionally private) that
crafts an e_lfanew value which wraps under uint arithmetic and asserts
the bounds check rejects it without any exception ever leaving the
method -- confirmed RED against the unpatched code (it threw
EndOfStreamException), GREEN after the fix.

+++ src/FalkForge.Engine.Protocol/Bundle/BundleReader.cs
@@ -341,7 +341,13 @@ public static class BundleReader
         if (ReadU16(stream, 0) != 0x5A4D) // 'MZ'
             return false;
 
-        var peOffset = ReadU32(stream, DosLfanewOffset);
+        // e_lfanew is a fully untrusted, attacker-controlled offset read straight from the DOS
+        // header. Cast to long immediately: peOffset + a const int operand otherwise binds to the
+        // (uint, uint) overload (the int operands are compile-time constants, which C# implicitly
+        // converts to uint), so a peOffset near uint.MaxValue would wrap the sum to a small value
+        // and defeat this bounds check. Every downstream use of peOffset below inherits this long
+        // type, so the same wrap cannot recur later in the method.
+        var peOffset = (long)ReadU32(stream, DosLfanewOffset);
         if (peOffset + PeSignatureSize + CoffHeaderSize > fileLength)
             return false;
         if (ReadU32(stream, peOffset) != 0x00004550) // 'P','E',0,0 (little-endian)
